requirements.txt
pyyaml==3.12
selenium==3.14.0
PyMySQL==0.9.2
peewee==3.7.1
requests==2.20.1
flask==1.0.2
xlrd==1.2.0
xlwt==1.3.0
ruamel.yaml==0.15.94
PyMySQL==0.9.2
openpyxl==2.6.2
sshtunnel==0.1.4